feat: Moved the Controller and Router to app folder
