+++ app/Controller/mainController.h
@@ -1,21 +1,20 @@
-// The ifndef is used so that the same File is not loaded again
-#ifndef INCLUDED_CONTROLLER
-#define INCLUDED_CONTROLLER
+#ifndef INCLUDED_MAIN_CONTROLLER
+#define INCLUDED_MAIN_CONTROLLER
 
-#include "../Request/Request.h"
-#include "../Response/Response.h"
+#include <iostream>
+
+#include "../../src/Logger/Logger.h"
+#include "../../src/Request/Request.h"
+#include "../../src/Response/Response.h"
 
 namespace Controller {
 void getHome(http::Request &request, http::Response &response);
 void getUser(http::Request &request, http::Response &response);
 void getUserChat(http::Request &request, http::Response &response);
 void getNotFound(http::Request &request, http::Response &response);
-void defaultNotFound(http::Request &request, http::Response &response);
-void getPublicFile(http::Request &request, http::Response &response);
 void saveUserData(http::Request &request, http::Response &response);
 void redirectUserToGoogle(http::Request &request, http::Response &response);
-
 void getXOGame(http::Request &request, http::Response &response);
-}  // namespace Controller
+};  // namespace Controller
 
 #endif
