+++ app/public/home/script.js
