+++ app/public/xo-game/script.js
