+++ app/public/xo-game/style.css
