+++ src/Router/Router.h
@@ -6,6 +6,7 @@
 #include <unordered_map>
 #include <vector>
 
+#include "../Controller/systemController.h"
 #include "../Request/Request.h"
 #include "../Response/Response.h"
